+++ module/apps/actorDetails.mjs
@@ -2,6 +2,7 @@ export class BRPactorDetails {
 
   // Get token/actor ID and type
   static async _getParticipantId(token, actor) {
+    console.log(token,actor)
     let particId = ""
     let particType = ""
     let particName = ""